horas-extras

+++ src/views/vsp/CreatePost.js
@@ -83,7 +83,10 @@ const CreatePost = () => {
 
         try {
             await store.collection('posts').add(post)
-            const { docs } = await store.collection('posts').get()
+            const { docs } = await store.collection('posts')
+            .orderBy('date', 'desc')
+            .where('autor', '==', 'administracion@vsp.com.co').get()
+        
             const nuevoArray = docs.map(item => ({ id: item.id, ...item.data() }))
             setPostUser(nuevoArray)
             console.log('post añadido')
@@ -102,7 +105,10 @@ const CreatePost = () => {
 
         try {
             await store.collection('posts').doc(id).delete()
-            const { docs } = await store.collection('posts').get()
+            const { docs } = await store.collection('posts')
+            .orderBy('date', 'desc')
+            .where('autor', '==', 'administracion@vsp.com.co').get()
+            
             const nuevoArray = docs.map(item => ({ id: item.id, ...item.data() }))
             setPostUser(nuevoArray)
         } catch (e) {
